--- src/stack.py
@@ -1,13 +1,0 @@
-class Stack:
-  def __init__(self):
-      self.size = 0
-      self.storage = list()
-
-  def push(self, value):
-      self.storage.append()
-
-  def pop(self):
-      pass
-
-  def len(self):
-      len(self.storage)
