+++ src/util.py
@@ -16,3 +16,17 @@ class Queue:
 
   def __len__(self):
     return self.len()
+
+class Stack:
+  def __init__(self):
+      self.size = 0
+      self.storage = list()
+
+  def push(self, value):
+      self.storage.append()
+
+  def pop(self):
+      pass
+
+  def len(self):
+      len(self.storage)
